botones de iniciar/cerrar sesion funcionando. falta funcionalidad boton user e intercambiar items
